test: move hero test to HomePage test file
